5. Cloud functions/requirements.txt
# Function dependencies, for example:
# package>=version
pandas
google-cloud-storage
pyarrow
fsspec
gcsfs
